Deliver a fat blow to YAML
YAML's biggest flaw is that it simply isn't a safe file format. A YAML file can be truncated virtually arbitrarily ... and remain a valid YAML file. This causes semantic corruption and missing elements in scenarios as routine as a partial file transfer, a nearly full filesystem, or an editor crash.  JSON, XML, everything short of dumb .ini files, has terminators and so avoids this problem.